[Mobile] Update BasicCalculator.
