+++ BasicCalculator/src/BasicCalculator/BasicCalculator.Tizen.Mobile/BasicCalculator/BasicCalculator.csproj
@@ -1,0 +1,12 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <!-- Include Nuget Package for Xamarin building -->
+  <ItemGroup>
+    <PackageReference Include="Tizen.NET" Version="11.0.0.17605" />
+  </ItemGroup>
+
+</Project>
