+++ BasicCalculator/src/BasicCalculator/BasicCalculator.Tizen.Mobile/Views/MobileMainView.xaml
@@ -2,7 +2,7 @@
 <View xmlns="http://tizen.org/Tizen.NUI/2018/XAML"
       xmlns:l="clr-namespace:BasicCalculator.Tizen.Mobile.Views"
       xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-      xmlns:viewModels="clr-namespace:BasicCalculator.ViewModels;assembly=BasicCalculator"
+      xmlns:viewModels="clr-namespace:BasicCalculator.ViewModels;assembly=BasicCalculator.Tizen.Mobile"
       x:Class="BasicCalculator.Tizen.Mobile.Views.MobileMainView">
 
     <View.XamlResources>
